Birdsong vs Background
This notebook trains a classifier on two different data sets.   The cornell bird song data set contains mp3 files varied in length, some are very long.   If you chop them into segments some will contain birdsongs but some clips will be just background sounds.     The goal is to have a classifier that can say "this clip contains a bird sound" without regard to the species.   This will be used as an upstream preprocessing step to decrease the incorrect labels in the downstream species classifier.

With roughly balanced classes, I get validated accuracy of about 90% before the validation loss increases (indicating overfitting).  

This was done with a vanilla CNN and very little tuneing.